requirements.txt
discord.py>=2.4.0
openai>=1.42.0
google-generativeai>=0.8.2
aiohttp>=3.9.5
aiosqlite>=0.20.0
python-dotenv>=1.0.1
pytest>=8.2.1
pytest-asyncio>=0.23.8
